+++ tests/Instances/Miscellaneous.hs
@@ -1,4 +1,4 @@
-﻿{-# LANGUAGE CPP                        #-}
+{-# LANGUAGE CPP                        #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE StandaloneDeriving         #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}